json: termv now merges the two json files.
The channels.json and streams.json are now combined
using their id and only the data required are extracted.

This takes about 2 secs to complete on my machine and is
triggered in the update function. Should investiagate on
how to reduce it.

Also added is sed as a dependency, need to change it to gawk for